Archive jwt based projects to enable uaa
This is because the current client container cannot connect to the gateway container. Future developments might explore the possibility. In a bid to make the gateway lighter it was necessary to move the UAA functionality into a several service. But the only to do that was to have workflows from version 6 which are incompatible to version 7 jhipster. So these will remain archived for now