Ammend persistence issues
The challenge seems to be that we are mutating data within an asynchronous workflow and this seems to really really upset the hibernate library. So we have simplified the workflow and the report system and the related persistence activities are working fine, except in one thing, that the workflow is now sysnchronizing the response which is something I find deeply undesirable. I would be remiss to not acknowldge that pulling off a jasper report service in the middle of a client response using AOP is pretty bad ass, but for pete's sake why is the asynchrnous executor not working, huh?